Backend/requirements.txt
sounddevice
numpy
torch
vosk
pyaudio
SpeechRecognition
edge_tts
pygame
pvporcupine
google-generativeai
psutil
pygetwindow
pyautogui
wmi
comtypes
pycaw